% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytome.R
\name{cytome_matrices}
\alias{cytome_matrices}
\title{Matrix catalogue (from `matrix_meta`)}
\usage{
cytome_matrices(x)
}
\arguments{
\item{x}{A `cytome` handle.}
}
\value{
A data.frame: matrix_name, n_rows, n_cols, dtype, row_entity, col_entity.
}
\description{
Matrix catalogue (from `matrix_meta`)
}
