% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytome.R
\name{cytome_obs}
\alias{cytome_obs}
\title{Read an entity table (cells / genes / peaks / tiles / samples) as a data.frame}
\usage{
cytome_obs(x, entity = "cells")
}
\arguments{
\item{x}{A `cytome` handle.}

\item{entity}{Entity table name. Default `"cells"` (obs). Use `"genes"`/`"peaks"` for var.}
}
\value{
A data.frame.
}
\description{
Read an entity table (cells / genes / peaks / tiles / samples) as a data.frame
}
